nexus-core/requirements.txt
fastapi
uvicorn[standard]
playwright
apscheduler
requests
python-dotenv